requirements.txt
langchain
# langchain-community
langchain-google-genai
ollama
python-dotenv
duckduckgo-search
wikipedia
arxiv

## multi-agent dependencies
crewai==0.22.5
crewai[tools]